% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_all.R
\name{bin_all}
\alias{bin_all}
\title{Bin All Features in a Dataframe}
\usage{
bin_all(dframe, y = "gb12", vars_to_exclude, vars_to_include, bins = NULL,
  tree_control = ctree_control(), verbose = FALSE)
}
\description{
Bin All Features in a Dataframe
}
